requirements.txt
pandas
requests
psycopg2-binary
feedparser
beautifulsoup4
altair